+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### ✅ Added
 - Added markdown support (enabled by default)
+- `LinkAttachmentView` and `LinkDetectionTextView` available for public use
 
 # [4.51.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.51.0)
 _March 26, 2024_